删除学生
批量删除学生

#### 添加学员

#### 根据id查询学生信息

#### 修改学生信息

#### 违纪处理相关

违纪分数 = 上一次 + 本次违纪扣分

违纪次数要加一

### 数据统计（可视化）

#### 学生学历统计

#### 班级人数统计

### 项目优化

当部门下有员工时则不允许删除该部门，并给前端提示错误信息：对不起，当前部门下有员工，不能直接删除！

+++ tlias-web-management/src/main/java/com/yince/service/impl/DeptServiceImpl.java
@@ -1,5 +1,6 @@
 package com.yince.service.impl;
 
+import com.yince.exception.KeyException;
 import com.yince.mapper.DeptMapper;
 import com.yince.pojo.Dept;
 import com.yince.service.DeptService;
@@ -36,6 +37,10 @@ public class DeptServiceImpl implements DeptService {
      */
     @Override
     public void deleteById(int id) {
+        // 1. 先查询部门下是否有员工
+        if (deptMapper.getEmpList(id) > 0) {
+            throw new KeyException(500,"约束异常");
+        }
         deptMapper.deleteById(id);
     }
